fix 3

+++ pages/send-transaction-confirmation/index.tsx
@@ -51,7 +51,7 @@ const SendTransactionConfirmation = () => {
     const connection = new Connection(process.env.NEXT_RPC_MAINNET_URL, "confirmed");
 
     console.log("Amount before: ", amount);
-    console.log("parsed amount: ", parseInt(amount));
+    console.log("parsed amount: ", parseFloat(amount));
     const transferParams = {
       connection,
       fromPubkey: new PublicKey(from),
